+++ pubspec.yaml
@@ -36,6 +36,7 @@ dependencies:
   flutter_staggered_grid_view: ^0.4.0
   firebase_core: ^1.3.0
   cloud_firestore: ^2.2.2
+  firebase_auth: ^3.0.1
   flutter_markdown: ^0.6.2
   
 dev_dependencies:
